LPS-61860 move to modules and use DS
